Fixed code to incorporate LED dimming (really Pulse-Width Modulation).
The Satellite schematic was altered to allow the use of the OCR0A pin, rather than a generic pin that requires the use of interrupts in order for the PWM to work. A second jumper was added that will allow for two devices to access both VCC and GND.

The code on the Satellite module was modularized for easy replication and/or reuse.  Work on the interrupt code was begun.  The interrupts will be used for the Zero-Cross Detection and for the external sensors used.