Mostly aesthetic changes to the PCB.
Added labels to all of the pins that may go off of the board, plus a few other helpful files.

There was a thin trace on the AC Hot output, so I fixed that up to a 0.032" width trace.

I found out that Sparkfun has a helpful schematic part called "Stand-Off".  It's based upon a 4-40 screw with the stand-off itself being about a quarter of an inch wide.  I replaced our old ones with this.