Lots of workarounds in the attempt to get AC and the PIR sensor working.  PIR sensor code is finalized and working.
Servo motor is not fully functional as of yet.  PTC Resettable fuse only goes up to 250mA, so we may be drawing too much for that.  PTC fuses will be ordered that support up to 2.7A (trip current).

Traces on PCB were ripped up in anticipation for alterations and re-trace.  Port C pins and VCC/GND pins were moved to the left side of the board in order to clear up space for the Bluetooth Module.  A ground plane will be implemented on the left side of the board (I feel the the right side is so entrenched in AC Neutral that pouring a ground plane will make little difference). 

I will also finish up with labels on all parts (especially capacitors). Final PCB design will be ordered tomorrow after our Thesis meeting.

Note: New files in this commit are a result of compiling the code in Release mode for testing purposes.