Added Interrupt Service Routines for both the PIR sensor and two for the Frequency Counter (one for the timer, one for the external interrupt).
Created a separate file for the Servo source code for maximum portability.  Also created header file to go along with it.

I also noticed that the files had put in my username of Dubs, so I changed it to my real name.

Changed Remote.sch to properly reflect the pins that are connected in the source code.

Moved footprints around in Satellite.sch so that there are no traces going underneath the Bluetooth module's antenna.

Moved all of the LINC files to a "Deprecated" folder to signify that it is no longer being used (RIP LINC).