Deprecated Remote code. RIP. Long live Android.
The Android code has been optimized visually for all platforms.  It is possible to change the SDK version of the project to 2.1 if we so decide (we can put in our paper that it will support over 95% of the Android population. http://developer.android.com/resources/dashboard/platform-versions.html
http://stackoverflow.com/questions/3614190/android-change-sdk-version-in-eclipse-unable-to-resolve-target-android-x

The LASAR Controller currently supports the ability to: save preferences, determine current Satellite settings for brightness and blinds-level, modify brightness and blinds-level in real-time, and has two MAC Addresses hard-coded into the program.

The Satellite now sends a ';' after all transmissions in order to aid the Android device in determining end-of-transmission.